attach '::timestamp' to named datetime placeholders
as of 1.22.0, pg8000 has dropped oid detection, and now instead
relies on postgres itself.. not sure if that was a good idea,
because postgres now has less information to work with..

in any case, this breaks the situation where we compare two
named datetime parameters, as in the added test. postgres decides
it cannot compare the two, because of missing timezone information
(no database column to match with).

pg8000 could perhaps pass the lack of timezone explicitly (will
file an issue), but for now we add ::timestamp to named datetime
parameters and that seems to work.