support storage=none
when using 'none' for the storage backend, the archive becomes
catalogue-only (similar to using --catalogue-only with the muninn
tools).

when ingesting a file, the remote_url property is set to the
path of the ingested file (file:///..).

not sure yet if we want to merge the tests for this with the main
archive tests.