Fix storing of double closing points in database for linearrings of polygons
Apparently we had a historic convention to use non-closed rings in muninn, but stored closed rings in the database.
To comply with the OGC standards all linearrings should be closed. We partially fixed this in 7.0.2, but this only impacted the interface through the muninn api. We still had the code in place to add a closing point when going to the database and remove this point when reading. This addition/removal step is now removed by this commit.